\name{indicators}
\alias{indicators}
\alias{protectedArea}
\title{ Test data sets for the eHab package }
\description{Data sets for resting and exampled in the eHab package}


\usage{data(eHab)}
\details{The indicators are some of the heavy metal concentrations of the
Meuse data set, interpolated to a grid. protectedArea is a random square
that is used to define the reference geometry.
}

\keyword{datasets}

