\name{eHab-package}
\alias{eHab-package}
\alias{eHab}
\docType{package}
\title{
Backend for e-Habitat
}
\description{
The package eHab will work as the computational backend for the environmental 
Web Processing Service e-Habitat.
}
\details{
\tabular{ll}{
Package: \tab eHab\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2010-06-04\cr
License: \tab GPL >= 2?\cr
LazyLoad: \tab yes\cr
}



}
\author{
Jon Olav Skoien
<jon.skoien@jrc.ec.europa.eu>
}

\keyword{ spatial}
\seealso{\link{pHabitat}, \link{hri}
}
